Support for Content Security Policy Header
Extends the CSP builder to support script-src and style-src, extends the Security Headers Provider to permit the injection of script-src and style-src directives, and implements hash generation for the on-the-fly JavaScript used for browser history and OIDC form redirect.

Adds script-src and style-src nonce support to Freemarker login templates with new NonceBean.

Signed-off-by: James Shuriff <james@febyte.com>